feat: partial native idle support
Client should now in theory idle when system is suspended on all
platforms, after 10 minutes on windows, mac, and xorg (I think?), and
when the lockscreen is triggered on windows and mac. Still need to work
on implementing wayland idle and lockscreen detection, most likely in a
native module.

Only tested with NixOS with hyprland, as expected getSystemIdleTime
doesn't work and neither does lock-screen or unlock-screen events, but
suspend and resume events work fine.